Move mod validators to separate namespace
