+++ src/Validator/Mod/SteamWorkshopArma3ModUrlValidator.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Validator;
+namespace App\Validator\Mod;
 
 use App\Service\SteamWorkshop\Exception\ItemNotFoundException;
 use App\Service\SteamWorkshop\Helper\Exception\InvalidItemUrlFormatException;
