+++ src/Validator/Mod/UniqueDirectoryModValidator.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Validator;
+namespace App\Validator\Mod;
 
 use App\Entity\Mod\DirectoryMod;
 use App\Form\Mod\Dto\ModFormDto;
