+++ src/Validator/Mod/WindowsDirectoryName.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Validator;
+namespace App\Validator\Mod;
 
 use Symfony\Component\Validator\Constraints\Regex;
